move MuiTheme into Layout from Index
